third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-css-3-expected.txt

Verifies CSS pretty-printing functionality.


Running: testMediaRule
====== 8< ------
@media screen,print {
    body {
        line-height: 1.2
    }
}

span {
    line-height: 10px
}

------ >8 ======
Correct mapping for <@media>
Correct mapping for <screen>
Correct mapping for <print>
Correct mapping for <body>
Correct mapping for <line-height>
Correct mapping for <1.2>

Running: testNamespaceRule
====== 8< ------
@namespace svg url(http://www.w3.org/2000/svg);g {
    color: red
}

------ >8 ======
Correct mapping for <namespace>
Correct mapping for <url>
Correct mapping for <color>
Correct mapping for <red>

Running: testPageRule
====== 8< ------
@page :first {
    margin: 2in 3in;
}

span {
    color: blue
}

------ >8 ======
Correct mapping for <page>
Correct mapping for <first>
Correct mapping for <margin>
Correct mapping for <3in>

Running: testSupportsRule
====== 8< ------
@supports(--foo:green) {
    body {
        color: green;
    }
}

#content {
    font-size: 14px
}

------ >8 ======
Correct mapping for <supports>
Correct mapping for <foo>
Correct mapping for <body>
Correct mapping for <color>

Running: testViewportRule
====== 8< ------
@viewport {
    zoom:0.75;min-zoom: 0.5;
    max-zoom: 0.9;
}

footer {
    position: fixed;
    bottom: 0;
}

------ >8 ======
Correct mapping for <viewport>
Correct mapping for <zoom>
Correct mapping for <0.5>
Correct mapping for <0.9>
